Simplify unregister
