+++ binding.gyp
@@ -37,8 +37,7 @@
         "native-src/virtual_drive/hydrate_file.cpp",
         "native-src/virtual_drive/register_sync_root/register_sync_root.cpp",
         "native-src/virtual_drive/register_sync_root/register_sync_root_wrapper.cpp",
-        "native-src/virtual_drive/unregister_sync_root/unregister_sync_root.cpp",
-        "native-src/virtual_drive/unregister_sync_root/unregister_sync_root_wrapper.cpp"
+        "native-src/virtual_drive/unregister_sync_root_wrapper.cpp"
       ],
       "include_dirs": [
         "include",
@@ -48,8 +47,7 @@
         "include/sync_root_interface/callbacks",
         "include/virtual_drive",
         "include/virtual_drive/get_registered_sync_roots",
-        "include/virtual_drive/register_sync_root",
-        "include/virtual_drive/unregister_sync_root"
+        "include/virtual_drive/register_sync_root"
       ],
       "libraries": [
         "-lCldApi.lib",
