--- include/virtual_drive/unregister_sync_root/unregister_sync_root.h
@@ -1,6 +1,0 @@
-#pragma once
-
-#include <node_api.h>
-
-HRESULT unregister_sync_root(const GUID &providerId);
-HRESULT unregister_sync_root(const wchar_t *providerIdStr);
