+++ include/virtual_drive/unregister_sync_root_wrapper.h
